+++ packages/json-rpc-engine/src/v2/index.test.ts
@@ -9,6 +9,7 @@ describe('@metamask/json-rpc-engine/v2', () => {
         "JsonRpcServer",
         "MiddlewareContext",
         "asLegacyMiddleware",
+        "createOriginMiddleware",
         "createScaffoldMiddleware",
         "getUniqueId",
         "isNotification",